feat: use lambda function
