+++ .projen/tasks.json
@@ -88,6 +88,10 @@
       "description": "Only compile",
       "steps": [
         {
+          "exec": "npm ci && npm run build",
+          "cwd": "lambda"
+        },
+        {
           "exec": "jsii --silence-warnings=reserved-word"
         }
       ]
