+++ lambda/util/iot.ts
@@ -1,0 +1,3 @@
+export const getCertIdFromARN = (arn: string) => {
+  return arn.split('/')[1];
+};
