+++ package.json
@@ -36,14 +36,14 @@
     "organization": false
   },
   "devDependencies": {
-    "@aws-cdk/integ-runner": "2.100.0-alpha.0",
-    "@aws-cdk/integ-tests-alpha": "2.100.0-alpha.0",
+    "@aws-cdk/integ-runner": "2.150.0-alpha.0",
+    "@aws-cdk/integ-tests-alpha": "2.150.0-alpha.0",
     "@stylistic/eslint-plugin": "^2",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.10.0",
     "@typescript-eslint/eslint-plugin": "^8",
     "@typescript-eslint/parser": "^8",
-    "aws-cdk-lib": "2.100.0",
+    "aws-cdk-lib": "2.150.0",
     "commit-and-tag-version": "^12",
     "constructs": "10.0.5",
     "eslint": "^9",
@@ -62,7 +62,7 @@
     "typescript": "^5.7.2"
   },
   "peerDependencies": {
-    "aws-cdk-lib": "^2.100.0",
+    "aws-cdk-lib": "^2.150.0",
     "constructs": "^10.0.5"
   },
   "dependencies": {
@@ -149,5 +149,6 @@
       "rootDir": "src"
     }
   },
-  "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
+  "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\".",
+  "packageManager": "yarn@1.22.21+sha1.1959a18351b811cdeedbd484a8f86c3cc3bbaf72"
 }
