--- src/lambda/util/iot.ts
@@ -1,3 +1,0 @@
-export const getCertIdFromARN = (arn: string) => {
-  return arn.split('/')[1];
-};
