+++ test/integ.thing-with-cert.js.snapshot/cdk.out
@@ -1,1 +1,1 @@
-{"version":"34.0.0"}
+{"version":"36.0.0"}
